requirements.txt
Adafruit_IO==2.7.1
keras==2.15.0
numpy==1.26.4
opencv_python==4.9.0.80
paho-mqtt==1.6.1
pyserial